+++ src/components/catalog-component.js
@@ -1,25 +1,34 @@
 import React from 'react'
-import PropTypes from 'prop-types'
 import withCatalog from './with-catalog'
 
+/**
+ * CatalogComponent is wrapped withCatalog by default and is capable of
+ *
+ * - rendering requested component from the catalog
+ * - passing props to the requested component
+ * - returns null (does not throw an error) when component does not exist in
+ *   the catalog
+ *
+ * Example:
+ *
+ * <CatalogComponent component="Button" hello="world" />
+ *
+ * CatalogComponent will now render the `<Button />` component (if it is part of
+ * the catalog and pass `hello` as a prop to the Button).
+ */
 class CatalogComponent extends React.Component {
-  static propTypes = {
-    catalog: PropTypes.object.isRequired,
-    component: PropTypes.string.isRequired,
-  }
-
   render() {
     const { catalog, component, ...others } = this.props
 
-    if (!catalog) {
-      return null
-    }
-    const Component = catalog.getComponent(component)
+    if (catalog && catalog.getComponent) {
+      const Component = catalog.getComponent(component)
 
-    if (!Component) {
-      return null
+      if (Component) {
+        return <Component {...others} />
+      }
     }
-    return <Component {...others} />
+
+    return null
   }
 }