+++ CHANGES.md
@@ -4,6 +4,16 @@ All notable changes to this project will be documented here. The format is based
 on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) and this project
 adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+### Changed
+
+- moved from old react Context API to latest [Context API](https://reactjs.org/docs/context.html)
+
+### Removed
+
+- `prop-types` dependency
+
 ## 2019/02/01 0.2.0
 
 ### Added